vm+compiler: native promises and async/await (#1319 phase 1)
Adds coroutine-style asynchronous programming to LPC: a first-class
promise value type, an `async` function modifier, an `await` expression
with real VM frame suspension, and `acatch`, the async-aware `catch`.
Cap'n Proto RPC, the `remote` modifier and transports from the RFC are
NOT included; function-flag bit 0x0080 is reserved for `remote`.

Value type
- T_PROMISE (0x10000, the first tag above the full 16-bit range), a
  ref-counted svalue in T_REFED, dispatched at every type site: free /
  assign, equality and identity, mapping keys, %O printing, typeof,
  save_object (ignored, like objects), sizing, reclaim, the cycle
  walkers, and the debug ref-checker.
- `promise` is also a declared compile-time type -- variables,
  parameters, return types and arrays -- so a statically-known misuse is
  a compile error. Calling an async function types as `promise` on both
  local-call paths.
- efuns: promise_create/resolve/reject/then/catch/status/result.
  Settlement is never synchronous: reactions are delivered from a
  microtask drain within the current gametick, each with a fresh
  eval-cost budget. Resolving with a promise adopts it (flattening);
  first settle wins; unhandled rejections are reported at deallocation.

Coroutines
- `async` (FUNC_ASYNC) propagates through inheritance and applies on
  every call path: local, `::`-qualified, call_other, funptrs, and
  simul_efuns. The body runs eagerly to its first await; the caller
  always receives a promise.
- `await` on a promise ALWAYS parks -- even an already-settled one -- so
  each step resumes with a fresh eval budget and a chain of awaits
  breaks long work into separately-metered pieces rather than burning
  one budget until "too long evaluation". Awaiting a non-promise passes
  through: a plain value is not a scheduling point.
- Parking copies the single frame (value-stack slice, resume pc, defers,
  acatch markers) into a coroutine owned by the awaited promise;
  settlement rebuilds it at top level. Destructed or recompiled owners
  reject cleanly; slices holding transient lvalues refuse to suspend
  with a clear error; the number of parked frames is bounded by the new
  `max suspended async functions` option (default 10000).
- `acatch(expr)` is a control-stack marker rather than a nested
  interpreter, so `await` may suspend inside it; it catches both
  synchronous errors and rejections arriving long after the frame
  parked, and unwinds the control, command-giver and value stacks and
  the ref list exactly as restore_context() would.
- await_callout(handle) exposes a pending call_out as a promise
  (fulfilled with the callback's return value, rejected if removed);
  async_info() lists suspended frames, the call_out_info() counterpart.

Compiler
- Keywords async/await/acatch/promise; the `await` unary expression and
  the `acatch` construct (F_AWAIT / F_ACATCH / F_END_ACATCH appended to
  ops.spec); NODE_ACATCH through the optimizer, icode, --ast and the
  disassembler. Illegal contexts (await outside async, inside catch, or
  inside a functional; acatch outside async or inside catch; async on a
  variable or in a modifier block) are compile errors.
- Regenerated grammar.autogen.* and the lpc-syntax grammar artifacts.

Testing
- tests/efuns/promise.lpc, tests/efuns/async_await.lpc,
  tests/operators/await.lpc, tests/compiler/async_await_context.lpc and
  eleven compiler/fail fixtures cover the value type, the scheduling
  model, acatch, await_callout, async_info and every legality rule.
  Deferred (post-resume) assertions are gated to fail the run and were
  probe-verified to do so; the memory-safety regressions were verified
  to fail on the unfixed binary.
- Validated on Debug (check_memory ref-count gate clean) and
  RelWithDebInfo, repeated randomized-order runs, plus GTest, the
  lpc-syntax tests and the corpus format check.

Docs: docs/concepts/general/async.md is the execution-model spec;
docs/efun/promises/ documents each efun; config.md and the sidebar are
regenerated.

+++ src/compiler/internal/keyword.h
@@ -14,13 +14,14 @@ struct keyword_t {
   unsigned short sem_value; /* semantic value for predefined tokens */
   struct ident_hash_elem_t* next;
   /* the fields above must correspond to struct ident_hash_elem */
-  short min_args;           /* Minimum number of arguments. */
-  short max_args;           /* Maximum number of arguments. */
-  unsigned short ret_type;  /* The return type used by the compiler. */
-  unsigned short arg_type1; /* Type of argument 1 */
-  unsigned short arg_type2; /* Type of argument 2 */
-  unsigned short arg_type3; /* Type of argument 3 */
-  unsigned short arg_type4; /* Type of argument 4 */
+  short min_args;          /* Minimum number of arguments. */
+  short max_args;          /* Maximum number of arguments. */
+  unsigned short ret_type; /* The return type used by the compiler. */
+  /* runtime T_* masks: 32-bit, the 16 low bits are fully allocated */
+  uint32_t arg_type1; /* Type of argument 1 */
+  uint32_t arg_type2; /* Type of argument 2 */
+  uint32_t arg_type3; /* Type of argument 3 */
+  uint32_t arg_type4; /* Type of argument 4 */
   short arg_index;          /* Index pointing to where to find arg type */
   short Default;            /* an efun to use as default for last
                              * argument */
